socrates-agent/requirements.txt
# Core AI
torch==2.8.0
torchaudio==2.8.0
transformers==4.56.0
tokenizers==0.22.0
sentence-transformers==5.1.0
suno-bark==0.0.1a0
llama_cpp_python==0.3.16

# LangChain & LangGraph
langchain==0.3.27
langsmith==0.4.21
langgraph==0.6.6

# Wikipedia API
wikipedia==1.4.0

# Database
psycopg2-binary==2.9.10

# Utilities
requests==2.32.5
numpy==2.3.2